refactor: abstracted the onloadui to smaller components
