+++ frontend/src/MainPage/OnLoadUIFolder/ChatGPTPlusLogo.tsx
@@ -1,0 +1,11 @@
+
+export const ChatGPTPlusLogo = () => {
+  return (
+    <h1 className="text-4xl font-semibold text-center text-gray-600 dark:text-200 dark:gray-600 ml-auto mr-auto mb-10 flex gap-2 items-center justify-center flex-grow">
+      ChatGPT
+      <span className="bg-yellow-200 text-yellow-900 py-0.5 px-1.5 text-xs md:text-sm rounded-md uppercase">
+        Plus
+      </span>
+    </h1>
+  )
+}
